requirements/linux-py3.8-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=linux-py3.8-requirements.txt
#
jinja2==3.1.3
    # via -r requirements.in
markupsafe==2.1.4
    # via jinja2
mysql-connector-python==8.2.0
    # via -r requirements.in
protobuf==4.21.12
    # via mysql-connector-python
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0.1
    # via -r requirements.in
tqdm==4.66.1
    # via -r requirements.in
